requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
gunicorn==21.2.0
